Check for existence of site file.

+++ checksites.sh
@@ -5,6 +5,11 @@
 
 SITE_LIST="site_list.txt"
 
+if [[ ! -f $SITE_LIST ]];then
+  echo "No site_list.txt file available! Please create one first."
+  exit 1
+fi
+
 while read url; do
   RESPONSE=`curl -L -s -o /dev/null -w "%{http_code}" ${url}`
   if [[ $RESPONSE == 401 ]]; then
